build: split acinclude.m4
The automake manual recommends against acinclude.m4:

"The first possibility (the historical practice) is to list
all your macros in `acinclude.m4'. This file will be included
in `aclocal.m4' when you run `aclocal', and its macro(s) will
henceforth be visible to `autoconf'. However if it contains
numerous macros, it will rapidly become difficult to maintain,
and it will be almost impossible to share macros between
packages."

Signed-off-by: Laszlo Papp <djszapi@archlinux.us>